feat: 카테고리탭, 검색 UI 추가
