+++ src/styles/globals.css
@@ -99,6 +99,7 @@
     --faq-list-a-padding-h: 0;
     --faq-list-a-size: 16px;
     --faq-list-q-padding: 16px 0;
+    --clear-space: 8px;
   }
   @media (min-width: 744px) {
     :root {
@@ -159,6 +160,7 @@
       --faq-list-a-padding-h: 24px;
       --faq-list-a-size: 20px;
       --faq-list-q-padding: 24px 24px;
+      --clear-space: 16px;
     }
   }
   @media (min-width: 1024px) {
